client-library/distexgpu/client.py
# This file provides a high-level interface for interacting with the central server.

class Client:
    def __init__(self, api_key: str):
        """
        Initializes the client with the given API key.
        
        :param api_key: A string representing the API key for authentication.
        """
        self.api_key = api_key

    def submit_job(self, gpu_requirements: int, input_data: dict) -> dict:
        """
        Submits a job to the central server.
        
        :param gpu_requirements: An integer representing the number of GPUs required for the job.
        :param input_data: A dictionary containing the input data for the job.
        :return: A dictionary containing the job submission response.
        """
        # Pseudo code for job submission
        return {"job_id": "12345", "status": "submitted"}

    def check_job_status(self, job_id: str) -> dict:
        """
        Checks the status of a submitted job.
        
        :param job_id: A string representing the unique ID of the job.
        :return: A dictionary containing the job status.
        """
        # Pseudo code for checking job status
        return {"job_id": job_id, "status": "completed"}

    def retrieve_results(self, job_id: str) -> dict:
        """
        Retrieves the results of a completed job.
        
        :param job_id: A string representing the unique ID of the job.
        :return: A dictionary containing the job results.
        """
        # Pseudo code for retrieving job results
        return {"job_id": job_id, "results": {"output": "data"}}
